aggregate all reference resolver to avoid cycle dependency
